Fix: common/optional: unaligned access when setting value
clang warns that setting the watchdog_timer value may result
in an unaligned access withing the LTTNG_OPTIONAL_SET macro.

To work around that, the macro invokes two functions which
defer to memset/memcpy (details::lttng_optional_set), one of
which is overloaded for r-values which require a local copy
to be performed before invoking memcpy.

+++ src/common/channel.cpp
@@ -271,8 +271,9 @@ void lttng_channel_set_default_extended_attr(struct lttng_domain *domain,
 		case LTTNG_BUFFER_PER_UID:
 			extended_attr->monitor_timer_interval =
 				DEFAULT_UST_UID_CHANNEL_MONITOR_TIMER;
-			LTTNG_OPTIONAL_SET(&extended_attr->watchdog_timer_interval,
-					   DEFAULT_UST_UID_CHANNEL_WATCHDOG_TIMER);
+			LTTNG_OPTIONAL_SET(
+				&extended_attr->watchdog_timer_interval,
+				static_cast<std::uint64_t>(DEFAULT_UST_UID_CHANNEL_WATCHDOG_TIMER));
 			extended_attr->blocking_timeout = DEFAULT_UST_UID_CHANNEL_BLOCKING_TIMEOUT;
 			break;
 		case LTTNG_BUFFER_PER_PID:
